requirements.txt
# OpenClaw Agent Dependencies
# ============================

# Core
pydantic>=2.0.0
pyyaml>=6.0
python-dotenv>=1.0.0
rich>=13.0.0
typer>=0.9.0

# LLM Providers
anthropic>=0.40.0
google-generativeai>=0.8.0

# Web Search
duckduckgo-search>=6.0.0
httpx>=0.27.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
youtube-transcript-api>=0.6.0

# Optional: Premium Search APIs
# serpapi>=0.1.5
# tavily-python>=0.3.0

# Async Support
aiohttp>=3.9.0
aiofiles>=23.0.0

# Web Server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
websockets>=12.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0

# Development
black>=24.0.0
ruff>=0.1.0
